Docs : move style components
